// src/services/idexpress/dto/api-create-order.dto.ts
export class PayloadOrderDto {
  orderNo: string;
  waybillNo?: string;
  orderTime: number;
  expressType: string;
  itemName: string;
  insured?: string;
  itemRemarks?: string;
  itemQuantity: number;
  itemCategory: string;
  weight: string;
  length?: number;
  width?: number;
  height?: number;
  serviceType: string;
  itemValue?: string;
  senderName: string;
  senderEmail?: string;
  senderPhoneNumber?: string;
  senderCellphone: string;
  senderProvinceId?: number;
  senderCityId: number;
  senderDistrictId: number;
  senderAddress: string;
  senderZipCode?: string;
  recipientName: string;
  recipientEmail?: string;
  recipientPhoneNumber?: string;
  recipientCellphone: string;
  recipientProvinceId?: number;
  recipientCityId: number;
  recipientDistrictId: number;
  recipientAddress: string;
  recipientZipCode?: string;
  codAmount?: string;
  paymentType: string;
  pickupStartTime: number;
  pickupEndTime: number;
}